AdminService.cs added, *change object to DTO*
