+++ PizzaApp/Controllers/AccountController.cs
@@ -10,74 +10,24 @@ namespace PizzaApp.Controllers;
 
 public class AccountController : BaseApiController
 {
-    private readonly UserManager<User> _userManager;
-    private readonly SignInManager<User> _signInManager;
-    private readonly ITokenService _tokenService;
+    private readonly IAccountService _accountService;
 
-    public AccountController(UserManager<User> userManager, SignInManager<User> signInManager,
-        ITokenService tokenService)
+    public AccountController(IAccountService accountService)
     {
-        _userManager = userManager;
-        _signInManager = signInManager;
-        _tokenService = tokenService;
+        _accountService = accountService;
     }
-
-
+    
     [HttpPost("register")]
     public async Task<ActionResult<UserDto>> Register(RegisterDto registerDto)
     {
-        if (await UserExists(registerDto.Email)) return BadRequest("Email is already in use");
-
-        var user = new User()
-        {
-            Email = registerDto.Email,
-            PhoneNumber = registerDto.PhoneNumber,
-            UserName = registerDto.Name,
-            Adress = registerDto.Adress
-        };
-
-        user.Email = registerDto.Email.ToLower();
-
-        var result = await _userManager.CreateAsync(user, registerDto.Password);
-        if (!result.Succeeded) return BadRequest(result.Errors);
-
-        var roleResult = await _userManager.AddToRoleAsync(user, "Customer");
-        if (!roleResult.Succeeded) return BadRequest(result.Errors);
-
-        return new UserDto()
-        {
-            Id = user.Id,
-            Email = registerDto.Email,
-            PhoneNumber = registerDto.PhoneNumber,
-            Name = registerDto.Name,
-            Adress = registerDto.Adress,
-            Token = await _tokenService.CreateToken(user)
-        };
+        //try catch for errors
+        return await _accountService.Register(registerDto);
     }
 
     [HttpPost("login")]
     public async Task<ActionResult<UserDto>> Login(LoginDto loginDto)
     {
-        var user = await _userManager.Users
-            .FirstOrDefaultAsync(x => x.Email.Equals(loginDto.Email.ToLower()));
-
-        if (user == null) return Unauthorized("Email doesn't exist");
-
-        var result = await _signInManager
-            .CheckPasswordSignInAsync(user, loginDto.Password, false);
-
-        if (!result.Succeeded) return Unauthorized();
-
-        return new UserDto()
-        {
-            Id = user.Id,
-            Name = user.UserName,
-            Token = await _tokenService.CreateToken(user),
-        };
-    }
-
-    private async Task<bool> UserExists(string username)
-    {
-        return await _userManager.Users.AnyAsync(x => x.Email.Equals(username.ToLower()));
+        //try catch for errors
+        return await _accountService.Login(loginDto);
     }
 }
