+++ PizzaApp/Controllers/AdminController.cs
@@ -10,48 +10,22 @@ namespace PizzaApp.Controllers;
 [Authorize(Roles = "Admin")]
 public class AdminController : BaseApiController
 {
-    private readonly IUnitOfWork _unitOfWork;
-    private readonly UserManager<User> _userManager;
+    private readonly IAdminService _adminService;
 
-    public AdminController(IUnitOfWork unitOfWork, UserManager<User> userManager)
+    public AdminController(IAdminService adminService)
     {
-        _unitOfWork = unitOfWork;
-        _userManager = userManager;
+        _adminService = adminService;
+       
     }
     [HttpGet]
     public async Task<ActionResult> GetUsersWithRoles()
     {
-        var users = await _userManager.Users
-            .Include(r => r.UserRoles)
-            .ThenInclude(r => r.Role)
-            .OrderBy(u => u.UserName)
-            .Select(u => new
-            {
-                u.Id,
-                Name = u.UserName,
-                Roles = u.UserRoles.Select(r => r.Role.Name).ToList()
-            })
-            .ToListAsync();
-
-        return Ok(users);
+        return Ok(await _adminService.GetUsersWithRoles());
     }
 
     [HttpPost("edit-roles/{userName}")]
     public async Task<ActionResult> EditRoles(string userName, [FromQuery] string roles)
     {
-        var selectedRoles = roles.Split(',').ToArray();
-        var user = await _userManager.FindByNameAsync(userName);
-
-        if (user == null) return NotFound("No user");
-
-        var userRoles = await _userManager.GetRolesAsync(user);
-
-        var result = await _userManager.AddToRolesAsync(user, selectedRoles.Except(userRoles));
-        if (!result.Succeeded) return BadRequest("Failed to add roles");
-
-        result = await _userManager.RemoveFromRolesAsync(user, userRoles.Except(selectedRoles));
-        if (!result.Succeeded) return BadRequest("Failed to remove roles");
-
-        return Ok(await _userManager.GetRolesAsync(user));
+        return Ok(await _adminService.EditRoles(userName, roles));
     }
 }
