+++ PizzaApp/Extensions/ApplicationServiceExtensions.cs
@@ -24,6 +24,8 @@ public static class ApplicationServiceExtensions
         services.AddScoped<IOrderService, OrderService>();
         services.AddScoped<IPizzaService, PizzaService>();
         services.AddScoped<ITopingService, TopingService>();
+        services.AddScoped<IUserService, UserService>();
+        services.AddScoped<IAccountService, AccountService>();
         services.AddSignalR();
         // services.AddSingleton<StateCheckerService>();
         var connectionString = configuration.GetConnectionString("DefaultConnection");
