+++ PizzaApp/Services/AdminService.cs
@@ -1,0 +1,54 @@
+﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using PizzaApp.Entities;
+using PizzaApp.Interfaces;
+
+namespace PizzaApp.Services;
+
+public class AdminService : IAdminService
+{
+    private readonly IUnitOfWork _unitOfWork;
+    private readonly UserManager<User> _userManager;
+
+    public AdminService(IUnitOfWork unitOfWork, UserManager<User> userManager)
+    {
+        _unitOfWork = unitOfWork;
+        _userManager = userManager;
+    }
+
+    public async Task<object> GetUsersWithRoles()
+    {
+        var users = await _userManager.Users
+            .Include(r => r.UserRoles)
+            .ThenInclude(r => r.Role)
+            .OrderBy(u => u.UserName)
+            .Select(u => new
+            {
+                u.Id,
+                Name = u.UserName,
+                Roles = u.UserRoles.Select(r => r.Role.Name).ToList()
+            })
+            .ToListAsync();
+
+        return users;
+    }
+
+    public async Task<object> EditRoles(string userName, string roles)
+    {
+        var selectedRoles = roles.Split(',').ToArray();
+        var user = await _userManager.FindByNameAsync(userName);
+
+        if (user == null) throw new  ApplicationException("No user");
+
+        var userRoles = await _userManager.GetRolesAsync(user);
+
+        var result = await _userManager.AddToRolesAsync(user, selectedRoles.Except(userRoles));
+        if (!result.Succeeded) throw new  ApplicationException("Failed to add roles");
+
+        result = await _userManager.RemoveFromRolesAsync(user, userRoles.Except(selectedRoles));
+        if (!result.Succeeded) throw new  ApplicationException("Failed to remove roles");
+
+        return await _userManager.GetRolesAsync(user);
+    }
+}
